CMaterialSet/CModel: Make use of size_t where applicable
Prevents type truncation warnings internally and also provides slightly
better code-gen, given the upper half of 64-bit registers don't need to
be constantly cleared.

+++ src/Core/Scene/CSceneNode.h
@@ -128,8 +128,8 @@ public:
     void LoadModelMatrix();
     void BuildLightList(CGameArea *pArea);
     void LoadLights(const SViewInfo& rkViewInfo);
-    void AddModelToRenderer(CRenderer *pRenderer, CModel *pModel, uint32 MatSet);
-    void DrawModelParts(CModel *pModel, FRenderOptions Options, uint32 MatSet, ERenderCommand RenderCommand);
+    void AddModelToRenderer(CRenderer *pRenderer, CModel *pModel, size_t MatSet);
+    void DrawModelParts(CModel *pModel, FRenderOptions Options, size_t MatSet, ERenderCommand RenderCommand);
     void DrawBoundingBox() const;
     void DrawRotationArrow() const;
